move keepass into wallet subdir
